requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cloudmersive-image-api-client==3.0.1
cloudmersive-nlp-api-client==4.0.1
cloudmersive-ocr-api-client==3.0.1
cloudmersive-validate-api-client==3.1.1
cryptography==3.3.1
cycler==0.10.0
Django==3.1.6
django-extensions==3.1.0
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-cloud==0.34.0
google-cloud-automl==2.1.0
google-cloud-secret-manager==2.2.0
google-cloud-vision==2.1.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.35.0
httplib2==0.18.1
idna==2.10
libcst==0.3.16
mypy-extensions==0.4.3
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.1
PyQtChart==5.12.0
PyQtWebEngine==5.12.1
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzbar==0.1.8
requests==2.25.1
rsa==4.7
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
wincertstore==0.2
